WIP: Current progress of Big Mars
All constants referencing max global parameters will have to be tweaked to use the GlobalParameterCalculator instead, so it can scale based off of the number of players.

BIG board is just a placeholder for testing, but eventually it will have to be implemented (probably using the same random setup as the full random board).